+++ app/Http/Controllers/Location/ZoneController.php
@@ -14,8 +14,12 @@ class ZoneController extends Controller
      */
     public function index()
     {
-        $zones = Zone::all();
-        return view('location.zone.index',compact('zones'));
+        if (\Auth::user()->can('manage-zone')) {
+            $zones = Zone::all();
+            return view('admin.location.zone.index', compact('zones'));
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
+        }
     }
 
     /**
@@ -23,7 +27,11 @@ class ZoneController extends Controller
      */
     public function create()
     {
-        return view('location.zone.create');
+        if (\Auth::user()->can('create-zone')) {
+            return view('admin.location.zone.create');
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
+        }
     }
 
     /**
@@ -31,15 +39,18 @@ class ZoneController extends Controller
      */
     public function store(Request $request)
     {
-        try{
-            $this->validate($request,[
-                'name' => 'required',
-            ]);
-            Zone::create($request->all());
-            return redirect()->back()->with('success', 'Zone Added Successfully.');
-        }catch (Exception $e)
-        {
-            return redirect()->back()->with('error', $e->getMessage());
+        if (\Auth::user()->can('create-zone')) {
+            try {
+                $this->validate($request, [
+                    'name' => 'required',
+                ]);
+                Zone::create($request->all());
+                return redirect()->route('admin.location.zone.index')->with('success', 'Zone Added Successfully.');
+            } catch (Exception $e) {
+                return redirect()->back()->with('error', $e->getMessage());
+            }
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
         }
     }
 
@@ -56,17 +67,25 @@ class ZoneController extends Controller
      */
     public function edit(Zone $zone)
     {
-        return view('location.zone.edit', compact('zone'));
+        if (\Auth::user()->can('edit-zone')) {
+            return view('admin.location.zone.edit', compact('zone'));
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
+        }
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request,$id)
+    public function update(Request $request, $id)
     {
-        $zone = Zone::find($id);
-        $zone->update($request->all());
-        return redirect()->back()->with('success', 'Zone Updated Successfully.'); 
+        if (\Auth::user()->can('edit-zone')) {
+            $zone = Zone::find($id);
+            $zone->update($request->all());
+            return redirect()->route('admin.location.zone.index')->with('success', 'Zone Updated Successfully.');
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
+        }
     }
 
     /**
@@ -74,8 +93,12 @@ class ZoneController extends Controller
      */
     public function destroy($id)
     {
-        $zone = Zone::find($id);
-        $zone->delete();
-        return redirect()->back()->with('success', 'Zone Deleted Successfully.');
+        if (\Auth::user()->can('delete-zone')) {
+            $zone = Zone::find($id);
+            $zone->delete();
+            return redirect()->back()->with('success', 'Zone Deleted Successfully.');
+        } else {
+            return redirect()->back()->with('error', __('Permission denied.'));
+        }
     }
 }
